cgroups are a decidedly Linux-only feature.  Build and test as such.
