+++ client/fingerprint/cgroup.go
@@ -1,11 +1,9 @@
 package fingerprint
 
 import (
-	"fmt"
 	"log"
 	"time"
 
-	client "github.com/hashicorp/nomad/client/config"
 	"github.com/hashicorp/nomad/nomad/structs"
 )
 
@@ -45,35 +43,6 @@ func NewCGroupFingerprint(logger *log.Logger) Fingerprint {
 		mountPointDetector: &DefaultMountPointDetector{},
 	}
 	return f
-}
-
-// Fingerprint tries to find a valid cgroup moint point
-func (f *CGroupFingerprint) Fingerprint(cfg *client.Config, node *structs.Node) (bool, error) {
-	mount, err := f.mountPointDetector.MountPoint()
-	if err != nil {
-		f.clearCGroupAttributes(node)
-		return false, fmt.Errorf("Failed to discover cgroup mount point: %s", err)
-	}
-
-	// Check if a cgroup mount point was found
-	if mount == "" {
-		// Clear any attributes from the previous fingerprint.
-		f.clearCGroupAttributes(node)
-
-		if f.lastState == cgroupAvailable {
-			f.logger.Printf("[INFO] fingerprint.cgroups: cgroups are unavailable")
-		}
-		f.lastState = cgroupUnavailable
-		return true, nil
-	}
-
-	node.Attributes["unique.cgroup.mountpoint"] = mount
-
-	if f.lastState == cgroupUnavailable {
-		f.logger.Printf("[INFO] fingerprint.cgroups: cgroups are available")
-	}
-	f.lastState = cgroupAvailable
-	return true, nil
 }
 
 // clearCGroupAttributes clears any node attributes related to cgroups that might
