+++ client/fingerprint/cgroup_linux.go
@@ -3,6 +3,10 @@
 package fingerprint
 
 import (
+	"fmt"
+
+	client "github.com/hashicorp/nomad/client/config"
+	"github.com/hashicorp/nomad/nomad/structs"
 	"github.com/opencontainers/runc/libcontainer/cgroups"
 )
 
@@ -21,4 +25,33 @@ func FindCgroupMountpointDir() (string, error) {
 		}
 	}
 	return mount, nil
+}
+
+// Fingerprint tries to find a valid cgroup moint point
+func (f *CGroupFingerprint) Fingerprint(cfg *client.Config, node *structs.Node) (bool, error) {
+	mount, err := f.mountPointDetector.MountPoint()
+	if err != nil {
+		f.clearCGroupAttributes(node)
+		return false, fmt.Errorf("Failed to discover cgroup mount point: %s", err)
+	}
+
+	// Check if a cgroup mount point was found
+	if mount == "" {
+		// Clear any attributes from the previous fingerprint.
+		f.clearCGroupAttributes(node)
+
+		if f.lastState == cgroupAvailable {
+			f.logger.Printf("[INFO] fingerprint.cgroups: cgroups are unavailable")
+		}
+		f.lastState = cgroupUnavailable
+		return true, nil
+	}
+
+	node.Attributes["unique.cgroup.mountpoint"] = mount
+
+	if f.lastState == cgroupUnavailable {
+		f.logger.Printf("[INFO] fingerprint.cgroups: cgroups are available")
+	}
+	f.lastState = cgroupAvailable
+	return true, nil
 }
